refactor: 관리자 및 대시보드 관련 타입을 별도 파일로 분리
